Move RunOptions and Context::try_as_run_options
